Added deconstructors to VectorPool
VectorPool was using placement new to initialize new objects, but
was never calling the destructor.  It didn't matter for a while,
but now that we're starting to include complicated objects in our
structures, this was starting to cause problems.  (This way is
also more correct.)

Also corrected an unrelated spelling error in comments.

Change-Id: I144c38e42bbd30a1497e8f78ddebd1f40ee69f6a