Add a way to specify the layer used by the raycast
Adding an optional parameter to RaycastSingle to specify the layer used
by the raycast, instead of the default one. Note that the object has to
collide with the layer being used, it does not go against the collision
type of the shape.

Tested on Linux, by using the editor to select against non-raycastable
objects.

Change-Id: I36e750394069887372021c5770632f34f4d9d80c