Fix physics collision checks going out of bounds
Collisions with the mouth would cause patrons to fall, disabling their
physics. This would happen while iterating over the number of manifolds
and contact points, which was occasionally causing an assert in Bullet
to fire because of using a no longer valid index.

Tested on Linux, Android, and Mac OSX

Change-Id: I3f3eb453341d03bc660f6d2ee21a743a4000172d