Create AnimationComponent.
Apply animation to the root node only, for now.

Going forward, we'll set uniforms for an array of matrices,
each matrix corresponding to the position of a bone.
The positions of the bones come from AnimationData, which
will be upgraded to hold a RigMotivator (which animates several
matrices), instead of a MatrixMotivator (which animates only
one matrix).  Then, in the vertex shader, each vertex will
be repositioned to match its bone.

Tested: works on Linux and Android.
Change-Id: I88c413017ebdae5e496bb093f2fb54e7807e46b2