Including the transform scale into physics
Passing the scale information from the transform component into the
physics shape, using the built in local scaling of Bullet. This allows
the shape to always match the entity, even at various scales.

Tested on Linux, using the editor to confirm that the sizes correctly
match the scale while changing.

Change-Id: I45b2e983fcc7999c4bcb4bcfe60ea20ec76fbf93