Removing the scale from physics during export.
The local scaling in physics modifies the half extents, radius, etc.
of shapes. When being exported, we want the unscaled size saved out,
as it reapplies the scale from the transform component on load. So,
removing the scale during the export process, to get the correct
size values.

Tested on Linux and Android.

Change-Id: I0aa96ebc6ba805b2e049935b3f876f5c3de3d741