Clearing physics data on add, fixes editor crash.
When an entity has a prototype with physics, and defines physics for
itself as well, it wasn't correctly clearing out the initial volumes
created for the prototype, which caused a crash within Bullet. Changing
it to clear the original data properly before clobbering it with local
changes.

Tested on Linux.

Change-Id: I65ab8c33ff0ce31a119c74aadf3cfc9417012191