Generating a static triangle mesh for physics.
Adding support for creating a btBvhTriangleMeshShape, which is a static
triangle mesh, as a shape to be added to an entity. This is done by
adding individual triangles to the mesh before hand, then generating
the base based on that. Note that concave shapes are supported.

These shapes are not intended to be added from flatbuffer format, but
instead added programatically. They are not written out by the editor.

Tested on Linux (in game and in editor) and Android.

Change-Id: Ie97a775b967d6cba11d59d63ce218ab5ac867b07