Fixed Windows build for VS2012.
This requires making some explicit move constructors and fixing a bunch of
things that Visual Studio is picky about (like not supporting brace
initializer lists for complex objects).

Also disabled some warnings that are treated as errors that cause the
build to fail.

This doesn't yet fully fix the build as there are glplatform.h errors that
remain to be fixed.

Change-Id: I429f9ea073ac98dbe77576d770f9dd5cef750f9a