Adding static redirect URLs for external linking.
Doxygen can do some pretty odd stuff with the URLs of the generated
HTML pages, so in order to allow simple linking from external
documentation (e.g. Scene Lab), I added the following fixed files.

The majority of them use simple javascript to check if the URL is
valid, and redirects appropriately; it redirects to the desired URL
if it is valid, otherwise it redirects to the fixed class-list page.

Note: There is no landing page for the 'component_library', so
'component_library.html' just uses an HTML <meta> tag refresh
to the class-list.

Tested: Tested using 'generate_docs.py' on Ubuntu and manually
        testing the redirects locally via Chrome.

Change-Id: I3af3be7b90687e36b54db3764bfca98cc2314a80