Hide Bullet physics implementation.
Hiding the implementation is nice because it isolates
the caller from all the Bullet physics header files.

The Bullet physics headers have some warnings
that are a pain to suppress everywhere physics.h
is used.

It's also just good coding practice to hide the
implementation, when possible.

Fixes OSX build, which was dying on Bullet physics header
warnings.

Tested: Works on Linux, Android, and OSX.
Change-Id: I606546567652938d0df974441273b9e1aa26601f