Propagating bone_transform changes to use mat4x3.
In order to save on the number of shader uniforms, the
bone_transforms now use `mathfu::Matrix<float, 4, 3>`, or
`typedef AffineTransform`.

Part of this change effects CORGI's `rendermesh`, so these files
were updated to use the new types instead of the traditional mat4.

Bug: b/24951383
Change-Id: I3df4ebcc18f3c839f6bff1781c33d1e95b201dd6
Tested: Tested by running Zooshi on Ubuntu.