Fixed up build for circular dependencies between CORGI and Breadboard.
The CORGI component_library build was failing when built as part of
Breadboard in the distribution source layout on Linux.  This was due
to Breadboard defining CORGI's location under dependencies/entity so
this change overrides this behavior by setting the CORGI location in
the project.

Also, fixed the standalone Android build (tested on Linux) as it simply
didn't work.

Tested:
Verified CORGI builds all targets (including Breadboard) on Linux
and for Android in the distribution source layout.

Bug: 22510457
Change-Id: I43fa429416c1cdc46e716b287f8820ffcff8e48f