Add Doxygen descriptors and turn off class diagrams.
Due to a Doxygen error, linklint fails when there is non-directive
text preceding a namespace. Currently the solution is to put a
brief descriptor above the namespaces until it's fixed.

Class diagrams also cause linklint to fail if the user does not
have graphviz installed, so they should be turned off by default.

Tested:
Works on OS X

Change-Id: I2bd5f43a5ef311e5128c4cca9e9a848f071d42ca