Fix duplicate entity id error logging.
The error handler is passed a reference to a string, which sometimes
points to the entity id that it is changing, causing the log to
repeat the new id twice, instead of listing the old id. Fixed by
caching the old id prior to the potential change.

Tested on Linux.

Change-Id: I03459ce1cf2be18e23070a5f867319f10b6c9c09