Refactor Shader interface for defines.
Shader reloading is now done explicitly with Shader::ReloadIfDirty(). The meaning of the dirty flag is clearer now (requires reload), and it can be explicitly set to force reload on source file update.

Explicit reloading allows the defines to be modified on any thread, and only reloaded on the Rendering thread, too.
PiperOrigin-RevId: 148108420
Change-Id: Ifccc375a28e98ad37532172d5578d42826e4b264